Update index.html
Hello, I would really appreciate if you could please make this change.  This change makes the image grow to fit the available space.  It may make really small images grainy which is why I coded it to be an option instead of always on, but most of the time I do not think those who use it will mind grainy images.  I tested it pretty well and it seems to work good.  The only part I did not test was whether or not it worked with the checkbox interaction because I did not know how to run it locally on my computer.    Thank you.  Please let me know if you want me to make any changes.